Bound host status token fallback
